Solved day 10b.
It's sloppy, but it works.
I need to write better libraries for working with points.
That would be the next iteration. I had to think about
points and such very deeply in order to get this one to
work. The debugging process was a pain in the butt as well.